Update patch set 1
Patch Set 1:

This needs a test otherwise we will break it in the future. I don't know how to test this, but hopefully we can devise something based on issue description.

I also concerned if this will slow everyone down. These files are not small, and (I imagine) it will take longer to copy the file than to move the file.

Alex

Patch-set: 1
CC: Gerrit User 5070 <5070@62eb7196-b449-3ce5-99f1-c037f21e1705>