Update patch set 1
Patch Set 1:

> This needs a test otherwise we will break it in the future. I don't know how to test this, but hopefully we can devise something based on issue description.

That's a fair point, but I also have no idea how to test it.

 
> I also concerned if this will slow everyone down. These files are not small, and (I imagine) it will take longer to copy the file than to move the file.

Yeah, it's not great.  It seems plausible to me that there is some way to test whether a copy is required, as there is on Unix systems, but I have no idea what it might be.

Patch-set: 1
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>