Update patch set 1
Patch Set 1:

I refreshed this CL but (after uncommenting the code as Elias pointed out) it breaks TestDotSlashImport and TestMultiplePackageImport which expect the shorter srcDir paths. I think we should probably leave this alone for now - the various clients that care about the problem have worked around it, and I'm fearful of breaking different clients by changing this.

The code is definitely going to be obsoleted as we push out package management support. Not in Go 1.10 but probably in Go 1.11. So at this point I think we should probably just leave this alone, not break any more clients, and wait for the new APIs, which hopefully will be much less fussy.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>