spec: match syntax for method expressions with implementations
A method expression is of the form T.m where T is a type and m
is a method of that type. The spec restricted T essentially to
a type name. Both cmd/compile and go/types accepted any type,
syntactically, and a method expression was really just a form
of a selector expression x.f where x denotes a type.

This CL removes the spec syntax restriction from MethodExpr
to match the actual implementation. It also moves MethodExpr
from Ooperand to PrimaryExpr, because that's what it is
syntactically in the implementation.

It still keeps the sepatate notion of MethodExpr, since it
must start with a type rather than a value, and the spec's
syntax expresses this bit of semantics via distinct productions
(e.g, conversions look like calls but also must start with a
type).

Fixes #9060.

Change-Id: Idd84655b5b4f85d7ee53ebf749f73f0414a05f4a