Update patch set 1
Patch Set 1:

I think that it might be better to discuss this stuff on golang-dev before moving forward, because #9661 is a long-standing issue. Some people want this for the sake of convenience, but it's still not clear to me what's the best API surface for representing the border of responsibility.

For example, Dial and Listen functions are complexes; Should the user-defined control function be applied to derived sockets such as DNS transport for dialing? Or should we have a separate control function for such stuff? Should we need to expose internal functionality that is now part of dialing/listening for supporting user-defined control functions?

Patch-set: 1
CC: Gerrit User 5076 <5076@62eb7196-b449-3ce5-99f1-c037f21e1705>