Update patch set 1
Patch Set 1:

> Thank you. Now I'm thinking about the significance of this CL. If user know the way to drop bit11 from modify FileHeader#Flags, This CL is not required.  One another way to do simplify this is adding function like FileHeader#ResetUTF8Flag(). But this will be exported from Reader too since this is in FileHeader. ResetUTF8Flag is meaningless for Reader.
> @dsnet How do you think?

How would ResetUTF8Flag() work? The logic added in CL 39570 activates no matter what. It won't know if the user intends to set the UTF-8 flag or not.

The situation with regard to ZIP encoding is awful (https://marcosc.com/2008/12/zip-files-and-encoding-i-hate-you/) and it is impossible to encode a ZIP file that is always compliant.

 * If we always enforce UTF-8, then we run the risk of several readers not properly having support for UTF-8.
 * If we only enforce UTF-8 when the strings are valid UTF-8 and is not compatible with CP-437 (which is our current policy), then we run this risk of several readers not properly interpreting the non-UTF-8 encoding as CP-437, but instead as whatever encoding they want.
 * If we only allow CP-437, then we disalllow a wide range of possible characters.

All situations are lose-lose-lose.

From the Go perspective, strings are conventionally UTF-8, but that is by no means a language enforced requirement. Thus, there is precedence that strings are UTF-8 and it is a reasonable default to assume Go strings are such, but it not required.

When Writer.WriteHeader is processing a FileHeader, there are three possible states the strings can be in:

 * DefinitelyUTF8: Name and Comment are definitely UTF-8; so set the UTF-8 flag. The generated file will be spec-compliant, but the user runs the risk of several readers not being able to read the file.
 * ProbablyUTF8: Name and Comment are possibly UTF-8; we don't know; so we use the detection logic added in CL 39570. However, this detection is not perfect. For example, it is possible for a GBK encoded string to still pass as valid UTF-8.
 * NotUTF8: Name and Comment are definitely not UTF-8. They are possibly in CP-437 or GBK or whatever. We just write the strings as is byte-for-byte. The generated file may not be spec-compliant, but at least it will work with the local variant of ZIP reader the user is targeting.

Since Go does not enforce UTF-8, every FileHeader starts off in the ProbablyUTF8 state, which will work most of the time, but will still unavoidably fail in some edge cases. For those edge cases, we need to provide a way for the user to encode their intent for the DefinitelyUTF8 or NotUTF8 states.

 * DefinitelyUTF8: Currently the user can achieve this by manually setting the UTF-8 bit in FileHeader.Flags. This is clunky, but at least it is possible and satisfies this use case.
 * NotUTF8: ???. We need to resolve this situation.

Thus, we can either add a single field FileHeader.NonUTF8 so the user can opt-in the the NotUTF8 state and continue requiring the user to manually set the UTF-8 bit for the DefinitelyUTF8 state.

Alternatively, we can add a FileHeader.Encoding field that is an enum where there zero-value is ProbablyUTF8, but allows the user to specify DefinitelyUTF8 or NotUTF8.

All solutions I can think of stink, but the reality is that ZIP file encodings is just absurd.

Patch-set: 1
Reviewer: Gerrit User 9735 <9735@62eb7196-b449-3ce5-99f1-c037f21e1705>