go/ast: add Apply, for rewriting ASTs
Credits: This is joint work of Josh Bleecher Snyder and I.
The code is based on a proposed implementation I had started
in #17108 and which was then reworked significantly by Josh
via golang.org/cl/55790.

This CL is essentially golang.org/cl/55790 with minor changes.
I expect to solidify/refine the API during the review cycle.

Fixes #17108.

Change-Id: Ibb3f9610f95814c3c56c70e228dd052f5cc9477b