go/ast: add Apply, for rewriting ASTs
Credits: This is joint work of Josh Bleecher Snyder and I.
The code is based on a proposed implementation I had started
in #17108 and which was then reworked significantly by Josh
via golang.org/cl/55790.

This CL is based on golang.org/cl/55790, with additional
input from Roger Peppe (reducing allocations), and some
API simplifications.

Fixes #17108.

Change-Id: Ibb3f9610f95814c3c56c70e228dd052f5cc9477b