runtime: make systemstack tail call if already switched
Currently systemstack always calls its argument, even if we're already
on the system stack. Unfortunately, traceback stops at the first
systemstack it sees, which often cuts off runtime stacks early,
leading to confusing panics and profiles.

Fix this on amd64 and 386 by performing a tail call if we're already
on the system stack. This eliminates it from the traceback entirely,
so it won't stop prematurely.

Change-Id: Ibc69e8765e899f8d3806078517b8c7314da196f4