time: document valid layouts are not valid Parse values
For #9346 #22135 explicitly state beside layout constants
that they are not valid time values for Parse. Also add 
examples of parsing valid RFC3339 values and the layout
to the example for time.Parse.

Fix capitalisation of time.Parse in Constants section.

For #20869 include RFC3339 in the list of layouts which do
not accept all the time formats allowed by RFCs (lowercase z).
This does not fully address #20869.

Fixes #9346 #22135

Change-Id: Ia4c13e5745de583db5ef7d5b1688d7768bc42c1b