go/printer: indent lone comments in composite lits
If a composite literal contains any comments on their own lines without
any elements, the printer would unindent the comments.

The comments in this edge case are written by the RBRACE print line
below - when the first argument is reached. Make the first argument be
an indent, so that they are printed with the needed indentation.

Fixes #22355.

Change-Id: I5ac25ac6de95a236f1e123479127cc4dd71e93fe