doc: move plain git examples in contribution guide
Plain git examples are littered throughout the contributing guidelines,
but are difficult to find when skim reading. This is largely because the
plain git examples are not included in code blocks like the
git-codereview code blocks. For "git aficionados" contributing will
appear less foreign if the plain git equivalents are more clearly
included alongside the git-codereview tools.

For screens wider than 500px the git-codereview and plain git examples
appear side-by-side, with the git-codereview example displayed on the
left and representing the recommended git-codereview command, and the
plain git example displayed on the right and representing the underlying
git commands that are run by git-codereview.

For screens up to and including 500px the git-codereview and plain git
examples appear one-after-the-other, with the git-codereview example
being displayed first.

Fixes #22484

Change-Id: Ic2c92db5724fc65959975a86e83a16f381496708