cmd/compile: do not try to write strings > 2g
Because ggloblsym takes lengths as int32, stringsym
converts sizes into int32 before calling it. If the
given string overflows an int32, then the size of the object
written into the obj file is negative and linking fails
later.

Add an error while writing the object file to give a
better explanation to the user what's wrong.

Fixes #1700.

Change-Id: I3933ce08ef846721ece7405bdba81dff644cb004