Update patch set 3
Patch Set 3:

(4 comments)

Did you consider using the xsmaxdp instruction? I'm not sure if it is guaranteed to be available, but if you can use it I think it would allow you to avoid the +/-0 checks (since the instruction handles those). Also, since it only matters if one of the operands is +inf if the other one is NaN I would push all the +inf handling code down into the NaN branch. In pseudocode something like:

    if isNaN(x) || isNaN(y) {
        goto nan_handler
    }
    return xsmaxdp(x, y)
nan_handler:
    // only need to materialize +inf if an operand is NaN
    // (which is very rare in real performance sensitive code)
    if isPosInf(x) || isPosInf(y) {
        return +inf
    }
    return nan

Something similar could also be done for Min using xsmindp. Even if you can't use the VSX instructions the code might be a bit faster if the +inf checks are taken out of the straight line path, and it might also be worth using FSEL to avoid a branch.

Patch-set: 3
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>