bytes: return slices where capacities don't overlap
The existing imlementations of Split, SplitN, SplitAfter, SplitAfterN,
Fields and FieldsFunc return slice views into the original backing array
with capacities that overlap. This can cause an append call on one of the
resulting slices to overflow into the next slice(s).

This change sets the capacities on the returned slice views such that
the views do not overlap. This makes sure that appending to one slice
has no effect on the others. If an append would overflow previously a
new backing array will now be allocated to fit the resulting slice.

Fixes #21149

Change-Id: I8934fb6daced2dec9b375ce77ea3de0e849d4f19