Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hmmm, I also just sent https://golang.org/cl/74510, sorry for the overlap.

Yeah I was about to respond to that CL just now (noticed it because of the merge issue). Is it alright for you if we keep this one of the two? It seems like a good issue for my first CL to Go.

Patch-set: 1
Reviewer: Gerrit User 23671 <23671@62eb7196-b449-3ce5-99f1-c037f21e1705>