Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > There is a slightly different approach in https://golang.org/cl/74530.
> 
> It seems cleaner to me to make the cap == len; it leads to a more predictable behavior when using append.

It's more predictable indeed, but it's more wasteful than necessary. Especially when the separators are made up of multiple bytes. Extending the capacity to include the following separator means a new allocation is not always needed when appending. 

Looking at this CL it seems like I forgot changing explode in mine btw, I'll fix that tomorrow. It would be nice if we can consolidate on a single thread to discuss things. For now, I'll go to bed though. I'll read responses tomorrow.

Patch-set: 2
CC: Gerrit User 23671 <23671@62eb7196-b449-3ce5-99f1-c037f21e1705>