Update patch set 1
Patch Set 1:

> At the cost of permitting accidental changes to the original slice.  The only point of this exercise is to minimize accidental changes, so I don't see a reason to prevent one kind while permitting another.

The original slice can still be changed by changing the views, even with a more tight cap. The only thing needed is to change a value at an existing index. The original issue was not that it changed the original slice (which is expected), it was that changes to one sub slice could overflow into another.

Patch-set: 1
Reviewer: Gerrit User 23671 <23671@62eb7196-b449-3ce5-99f1-c037f21e1705>