Update patch set 4
Patch Set 4:

> There's a change that this change will break existing code. If that's the case we want that code to break consistently. If sometimes there's extra space in the slice, we might miss those cases depending on input.

Alright, that is an argument I can get behind. Much more than wanting to limit changing the original slice, because I think the original slice should be considered consumed by Split. Mainly because changing bytes in the resulting views will still have effect on the original slice.

However a couple of things to take in consideration before deciding that this approach is the one we want:

- In the issue discussion, one of the main arguments against was performance loss.
- Once this approach has been chosen it will be impossible to change it to my approach, because my approach has weaker guaranteees, since the original slice can be changed when appending.

Patch-set: 4