Update patch set 1
Patch Set 1:

> > Yeah I was about to respond to that CL just now (noticed it because of the merge issue). Is it alright for you if we keep this one of the two? It seems like a good issue for my first CL to Go.
> 
> Well, I suppose it's OK, but I like my version better.  We don't need the slice caps to capture the separator; it's fine, and simpler, for them to just be the bytes of the field.  Want to change yours to look more like mine?


I commented on your CL as well with some thoughts on the difference in the fix. Let's discuss more on it there. If your approach is considered better in the end I think it's good if your CL is merged. Otherwise I'll clean up mine with some changes to the tests and fix for explode. Finally, next time please post on the issue that you're working on it, so we don't do double work.

Patch-set: 1
Reviewer: Gerrit User 23671 <23671@62eb7196-b449-3ce5-99f1-c037f21e1705>