cmd/dist: include timing info by default
Previously we added an environment variable to print timing
information to a file. This is helpful but it seems more helpful to
enable it for everyone. This has the side benefit of familiarizing all
people who compile from source with the average timing of each step in
the build process, and avoids the overhead of needing to write to
a log file.

Not logging from a common start point adds a possibility for
coordinated omission error, however it feels more useful to see the
step timings than the overall build timings.

Example command output:

    $ GOROOT_BOOTSTRAP=~/go1.9 ./make.bash
    Building Go cmd/dist using /Users/kevin/go1.9... 1s
    Building Go toolchain1 using /Users/kevin/go1.9... 9.5s
    Building Go bootstrap cmd/go (go_bootstrap) using Go toolchain1... 6.6s
    Building Go toolchain2 using go_bootstrap and Go toolchain1... 16.4s
    Building Go toolchain3 using go_bootstrap and Go toolchain2... 15.3s
    Building packages and commands for darwin/amd64... 16.4s
    ---
    Installed Go for darwin/amd64 in /Users/kevin/go
    Installed commands in /Users/kevin/go/bin

Updates #22532.

Change-Id: Ifa83ca8ea46f77de2bb5f09648afbf70f5205374