cmd/trace: implement focustask trace view mode
that is like the regular trace view, but highlights goroutine activities
related to the task specified as focustask and its subtasks.

Also, improved the logic to determine whether a goroutine execution
slice is related to the task by also considering user log events.
Previously, we classified a slice as the task-related only if there is
an overlapping 'span'. Now any trace annotation activities associated
with the task or the subtasks occurred on the goroutine at that time
will cause the slice to be highlighted.

Change-Id: I8adbd9fc3d09606957745c359e85b3dd0a1dcf78